+++ src/window/mod.rs
@@ -44,9 +44,11 @@ use timely::{Data, ExchangeData};
 
 pub(crate) mod clock;
 pub(crate) mod hopping_window;
+pub(crate) mod session_window;
 pub(crate) mod tumbling_window;
 
 use self::hopping_window::HoppingWindowConfig;
+use self::session_window::SessionWindow;
 use self::tumbling_window::TumblingWindowConfig;
 use clock::{event_time_clock::EventClockConfig, system_clock::SystemClockConfig, ClockConfig};
 
@@ -101,6 +103,8 @@ impl PyConfigClass<Box<dyn WindowBuilder>> for Py<WindowConfig> {
         if let Ok(conf) = self.extract::<TumblingWindowConfig>(py) {
             Ok(Box::new(conf))
         } else if let Ok(conf) = self.extract::<HoppingWindowConfig>(py) {
+            Ok(Box::new(conf))
+        } else if let Ok(conf) = self.extract::<SessionWindow>(py) {
             Ok(Box::new(conf))
         } else {
             let pytype = self.as_ref(py).get_type();
@@ -484,5 +488,6 @@ pub(crate) fn register(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<WindowConfig>()?;
     m.add_class::<TumblingWindowConfig>()?;
     m.add_class::<HoppingWindowConfig>()?;
+    m.add_class::<SessionWindow>()?;
     Ok(())
 }